+++ build.gradle
@@ -43,8 +43,8 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
-    mainClassName= "Launcher"
+    mainClass.set("URBOI_PACKIN")
+    mainClassName= "URBOI_PACKIN.Launcher"
 }
 
 shadowJar {
