+++ src/main/java/URBOI_PACKIN/Launcher.java
@@ -1,3 +1,5 @@
+package URBOI_PACKIN;
+
 import javafx.application.Application;
 
 /**
