+++ src/main/java/URBOI_PACKIN/Main.java
@@ -1,5 +1,8 @@
+package URBOI_PACKIN;
+
 import java.io.IOException;
 
+import URBOI_PACKIN.UI.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -7,11 +10,11 @@ import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
 /**
- * A GUI for Duke using FXML.
+ * A GUI for URBOI_PACKIN.ResponseController using FXML.
  */
 public class Main extends Application {
 
-    private Duke duke = new Duke();
+    private ResponseController responseController = new ResponseController();
 
     @Override
     public void start(Stage stage) {
@@ -20,7 +23,7 @@ public class Main extends Application {
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
+            fxmlLoader.<MainWindow>getController().setDuke(responseController);
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
