+++ src/main/java/URBOI_PACKIN/TaskTypes/Event.java
@@ -1,0 +1,27 @@
+package URBOI_PACKIN.TaskTypes;
+
+import URBOI_PACKIN.Task;
+
+/**
+ * Represents an Event task.
+ */
+public class Event extends Task {
+    protected String from;
+    protected String to;
+
+    public Event(String description, String from, String to) {
+        super(description, null);
+        this.from = from;
+        this.to = to;
+    }
+
+    @Override
+    public String toFileString() {
+        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from + " | " + to;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
+    }
+}
