+++ src/main/java/URBOI_PACKIN/UI/DialogBox.java
@@ -1,3 +1,5 @@
+package URBOI_PACKIN.UI;
+
 import java.io.IOException;
 import java.util.Collections;
